requirements.txt
asgiref==3.5.2
Django==4.0.5
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-htmx==1.12.1
django-js-asset==2.0.0
django-rest-framework==0.1.0
django-tinymce==3.4.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
django-ckeditor